完成前台用户模块

+++ guns/guns-admin/src/main/java/com/stylefeng/guns/modular/system/controller/UserApiController.java
@@ -1,6 +1,16 @@
 package com.stylefeng.guns.modular.system.controller;
 
+import com.baomidou.mybatisplus.mapper.EntityWrapper;
 import com.stylefeng.guns.core.base.controller.BaseController;
+import com.stylefeng.guns.core.common.exception.BizExceptionEnum;
+import com.stylefeng.guns.core.exception.GunsException;
+import com.stylefeng.guns.core.shiro.ShiroKit;
+import com.stylefeng.guns.core.support.DateTime;
+import com.stylefeng.guns.core.util.ToolUtil;
+import com.stylefeng.guns.modular.system.model.User;
+import com.stylefeng.guns.modular.system.model.UserInfo;
+import com.stylefeng.guns.modular.system.warpper.UserWorksWarpper;
+import com.stylefeng.guns.modular.userInfo.service.IUserInfoService;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -12,7 +22,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 import com.stylefeng.guns.modular.system.model.UserApi;
 import com.stylefeng.guns.modular.system.service.IUserApiService;
 
+import java.util.Date;
 import java.util.List;
+import java.util.Map;
 
 /**
  * 控制器
@@ -28,6 +40,9 @@ public class UserApiController extends BaseController {
 
     @Autowired
     private IUserApiService userApiService;
+
+    @Autowired
+    private IUserInfoService userInfoService;
 
     /**
      * 跳转到首页
@@ -62,7 +77,12 @@ public class UserApiController extends BaseController {
     @RequestMapping(value = "/list")
     @ResponseBody
     public Object list(String condition) {
-        return userApiService.selectList(null);
+
+        List<Map<String, Object>> list  = userApiService.list(condition);
+        return super.warpObject(new UserApiWarpper(list));
+
+
+
     }
 
     /**
@@ -71,17 +91,44 @@ public class UserApiController extends BaseController {
     @RequestMapping(value = "/add")
     @ResponseBody
     public Object add(UserApi userApi) {
+        userApi.setCreatetime(new DateTime());
         userApiService.insert(userApi);
+        UserInfo userInfo = new UserInfo();
+        userInfo.setUserId(userApi.getId());
+        userInfo.setApiToken(ShiroKit.getRandomSalt(5));
+        userInfo.setCredits(0);
+        userInfo.setMoney(0);
+        userInfo.setLoginIp("192.168.0.0.1");
+        userInfo.setCreateTime(new DateTime());
+        userInfo.setRealName(userApi.getName());
+        userInfo.setJoinClub(2);
+        userInfo.setAppointment(2);
+        userInfo.setEnlightening(2);
+        userInfoService.insert(userInfo);
+
         return SUCCESS_TIP;
     }
 
     /**
-     * 删除
+     * 前台用户删除
      */
     @RequestMapping(value = "/delete")
     @ResponseBody
     public Object delete(@RequestParam Integer userApiId) {
-        userApiService.deleteById(userApiId);
+        Boolean flag =false;
+        //先删除用户详情id
+        EntityWrapper<UserInfo> entityWrapper = new EntityWrapper<>();
+        entityWrapper.eq("user_id",userApiId);
+        List<UserInfo> userInfos = userInfoService.selectList(entityWrapper);
+        if(userInfos.size()>0){
+         flag =     userInfoService.deleteById(userInfos.get(0).getId());
+        }
+        if(flag){
+            userApiService.deleteById(userApiId);
+
+        }else {
+            return  "用户详情删除失败!";
+        }
         return SUCCESS_TIP;
     }
 
@@ -91,6 +138,7 @@ public class UserApiController extends BaseController {
     @RequestMapping(value = "/update")
     @ResponseBody
     public Object update(UserApi userApi) {
+        userApi.setUpdatetime(new DateTime());
         userApiService.updateById(userApi);
         return SUCCESS_TIP;
     }
@@ -115,5 +163,21 @@ public class UserApiController extends BaseController {
         return userApis;
     }
 
+    /**
+     * 跳转到用户详情页面页面
+     */
+    @RequestMapping("/user_info/{userId}")
+    public String userInfoDetail(@PathVariable Integer userId, Model model) {
+        if (ToolUtil.isEmpty(userId)) {
+            throw new GunsException(BizExceptionEnum.REQUEST_NULL);
+        }
+        UserApi user = userApiService.selectById(userId);
+        EntityWrapper<UserInfo> userInfoEntityWrapper = new EntityWrapper<>();
+        userInfoEntityWrapper.eq("user_id", userId);
+        List<UserInfo> userInfos = userInfoService.selectList(userInfoEntityWrapper);
+        model.addAttribute("item", userInfos.get(0));
+        model.addAttribute("userName", user.getName());
+        return"/userInfo/userInfo/userInfo_edit.html";
+    }
 
 }
